test: added more tests for bentofileSchema.json; separated tests into one file per bentofile field
